Fix tests and update unity
