+++ Assets/Scripts/Utils/UtilsAssembly.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Behide.Utils",
+    "name": "UtilsAssembly",
     "rootNamespace": "",
     "references": [
         "GUID:6055be8ebefd69e48b49212b09b47b2f"
