+++ Assets/Tests/Common/Tests.Common.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "Tests.Common"
+}
