+++ Assets/Tests/IntegrationTests/InputSystemTests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63e0c8d2efb8cfd48ae9a6fbc85ff95a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
