+++ Assets/Tests/IntegrationTests/InputSystemTests/InputSystemTests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a76fb5e9f365c9542af0b809d7695e21
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
