+++ Assets/Tests/IntegrationTests/InputSystemTests/PhysicalButtonTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a209ee1aa97d7164db810e1210b15399
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
