+++ Assets/Tests/IntegrationTests/IntegrationnTests.asmdef
@@ -1,0 +1,25 @@
+{
+    "name": "IntegrationTests",
+    "rootNamespace": "",
+    "references": [
+        "GameAssembly",
+        "UtilsAssembly",
+        "Tests.Common",
+        "UnityEngine.TestRunner",
+        "Unity.InputSystem",
+        "Unity.InputSystem.TestFramework"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
