+++ Assets/Tests/IntegrationTests/Scene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 903cec3d7bd61b84f936320baaf4e393
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
