+++ Packages/manifest.json
@@ -1,18 +1,18 @@
 {
   "dependencies": {
     "com.unity.2d.sprite": "1.0.0",
-    "com.unity.ai.navigation": "1.1.3",
-    "com.unity.collab-proxy": "2.0.4",
+    "com.unity.ai.navigation": "1.1.4",
+    "com.unity.collab-proxy": "2.0.5",
     "com.unity.feature.development": "1.0.1",
     "com.unity.ide.visualstudio": "2.0.18",
     "com.unity.ide.vscode": "1.2.5",
-    "com.unity.inputsystem": "1.5.1",
+    "com.unity.inputsystem": "1.6.3",
     "com.unity.postprocessing": "3.2.2",
     "com.unity.remote-config": "3.3.2",
-    "com.unity.shadergraph": "14.0.7",
+    "com.unity.shadergraph": "14.0.8",
     "com.unity.test-framework": "1.1.33",
     "com.unity.textmeshpro": "3.0.6",
-    "com.unity.timeline": "1.7.4",
+    "com.unity.timeline": "1.7.5",
     "com.unity.toolchain.win-x86_64-linux-x86_64": "2.0.5",
     "com.unity.ugui": "1.0.0",
     "com.unity.vectorgraphics": "2.0.0-preview.21",
