+++ ProjectSettings/PackageManagerSettings.asset
@@ -12,12 +12,12 @@ MonoBehaviour:
   m_Script: {fileID: 13964, guid: 0000000000000000e000000000000000, type: 0}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_EnablePreReleasePackages: 1
+  m_EnablePreReleasePackages: 0
   m_AdvancedSettingsExpanded: 1
   m_ScopedRegistriesSettingsExpanded: 1
   m_SeeAllPackageVersions: 0
-  m_DismissPreviewPackagesInUse: 1
-  oneTimeWarningShown: 1
+  m_DismissPreviewPackagesInUse: 0
+  oneTimeWarningShown: 0
   m_Registries:
   - m_Id: main
     m_Name: 
@@ -31,6 +31,6 @@ MonoBehaviour:
   m_RegistryInfoDraft:
     m_Modified: 0
     m_ErrorMessage: 
-    m_UserModificationsInstanceId: -834
-    m_OriginalInstanceId: -836
+    m_UserModificationsInstanceId: -836
+    m_OriginalInstanceId: -838
   m_LoadAssets: 0
