+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2022.3.0f1
-m_EditorVersionWithRevision: 2022.3.0f1 (fb119bb0b476)
+m_EditorVersion: 2022.3.5f1
+m_EditorVersionWithRevision: 2022.3.5f1 (9674261d40ee)
